+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Fix misleading error message when a task parameter specified as a file path does not actually exist as a file.
+- Fix bug where if all possible dependency pathways are circularly dependent, this is not caught by MatFlow.
 
 ## [0.2.17] - 2021.02.15